Make float <-> int casts explicit outside of test, qt, CFeeRate
Summary:
Backport of core PR11303
Only the commit 53a6590 is applicable for us.
https://github.com/bitcoin/bitcoin/pull/11303/commits/53a6590f496b25174c740927243bf8307541b0b9

Test Plan:
  make check

Reviewers: #bitcoin_abc, deadalnix, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D3313

+++ src/wallet/wallet.cpp
@@ -392,9 +392,9 @@ bool CWallet::ChangeWalletPassphrase(
                                          pMasterKey.second.vchSalt,
                                          pMasterKey.second.nDeriveIterations,
                                          pMasterKey.second.nDerivationMethod);
-            pMasterKey.second.nDeriveIterations =
+            pMasterKey.second.nDeriveIterations = static_cast<unsigned int>(
                 pMasterKey.second.nDeriveIterations *
-                (100 / ((double)(GetTimeMillis() - nStartTime)));
+                (100 / ((double)(GetTimeMillis() - nStartTime))));
 
             nStartTime = GetTimeMillis();
             crypter.SetKeyFromPassphrase(strNewWalletPassphrase,
@@ -403,8 +403,9 @@ bool CWallet::ChangeWalletPassphrase(
                                          pMasterKey.second.nDerivationMethod);
             pMasterKey.second.nDeriveIterations =
                 (pMasterKey.second.nDeriveIterations +
-                 pMasterKey.second.nDeriveIterations * 100 /
-                     double(GetTimeMillis() - nStartTime)) /
+                 static_cast<unsigned int>(
+                     pMasterKey.second.nDeriveIterations * 100 /
+                     double(GetTimeMillis() - nStartTime))) /
                 2;
 
             if (pMasterKey.second.nDeriveIterations < 25000) {
@@ -678,8 +679,8 @@ bool CWallet::CreateMasteyKey(const SecureString &strWalletPassphrase,
     int64_t nStartTime = GetTimeMillis();
     crypter.SetKeyFromPassphrase(strWalletPassphrase, kMasterKey.vchSalt, 25000,
                                  kMasterKey.nDerivationMethod);
-    kMasterKey.nDeriveIterations =
-        2500000 / ((double)(GetTimeMillis() - nStartTime));
+    kMasterKey.nDeriveIterations = static_cast<unsigned int>(
+        2500000 / double(GetTimeMillis() - nStartTime));
 
     nStartTime = GetTimeMillis();
     crypter.SetKeyFromPassphrase(strWalletPassphrase, kMasterKey.vchSalt,
@@ -687,8 +688,8 @@ bool CWallet::CreateMasteyKey(const SecureString &strWalletPassphrase,
                                  kMasterKey.nDerivationMethod);
     kMasterKey.nDeriveIterations =
         (kMasterKey.nDeriveIterations +
-         kMasterKey.nDeriveIterations * 100 /
-             ((double)(GetTimeMillis() - nStartTime))) /
+         static_cast<unsigned int>(kMasterKey.nDeriveIterations * 100 /
+                                   double(GetTimeMillis() - nStartTime))) /
         2;
 
     if (kMasterKey.nDeriveIterations < 25000) {
